float to double

+++ common/coordinate.hpp
@@ -5,9 +5,9 @@
 #include <glm/ext.hpp>
 
 namespace rt {
-	inline glm::vec3 polar_to_cartesian(float theta, float phi) {
-		float sinTheta = std::sin(theta);
-		glm::vec3 v = {
+	inline glm::dvec3 polar_to_cartesian(double theta, double phi) {
+		double sinTheta = std::sin(theta);
+		glm::dvec3 v = {
 			sinTheta * std::cos(phi),
 			sinTheta * std::sin(phi),
 			cos(theta)
@@ -16,20 +16,20 @@ namespace rt {
 	};
 
 	// zが上の座標系に移動する行列
-	inline glm::mat3 to_bxdf_basis_transform(const glm::vec3 &n) {
-		glm::vec3 xaxis;
-		glm::vec3 zaxis = n;
-		glm::vec3 yaxis;
-		if (0.999f < glm::abs(zaxis.z)) {
-			xaxis = glm::normalize(glm::cross(glm::vec3(0.0f, 1.0f, 0.0f), zaxis));
+	inline glm::mat3 to_bxdf_basis_transform(const glm::dvec3 &n) {
+		glm::dvec3 xaxis;
+		glm::dvec3 zaxis = n;
+		glm::dvec3 yaxis;
+		if (0.999 < glm::abs(zaxis.z)) {
+			xaxis = glm::normalize(glm::cross(glm::dvec3(0.0, 1.0, 0.0), zaxis));
 		}
 		else {
-			xaxis = glm::normalize(glm::cross(glm::vec3(0.0f, 0.0f, 1.0f), zaxis));
+			xaxis = glm::normalize(glm::cross(glm::dvec3(0.0, 0.0, 1.0), zaxis));
 		}
 		yaxis = glm::cross(zaxis, xaxis);
 		return glm::transpose(glm::mat3(xaxis, yaxis, zaxis));
 	}
-	inline glm::vec3 from_bxdf(const glm::vec3 &n, const glm::vec3 &bxdf_dir) {
+	inline glm::dvec3 from_bxdf(const glm::dvec3 &n, const glm::dvec3 &bxdf_dir) {
 		return glm::transpose(to_bxdf_basis_transform(n)) * bxdf_dir;
 	}
